Improved data-set deletion functionality.
Added to data-set deletion functionality, including 2 new SQL scripts
for stored procedures. Altered the approveDataset method in the
DatasetService class so that child data-sets get marked for deletion
when their parent data-set is marked. Also added buttons to the app's UI
to execute deletion tasks. Added DatasetRowMapper class for further
implementation of JDBC in the future.